refactor(frontend): extract form out of SettingsPage into its own component
